requirements.txt
Flask==2.3.2
Flask-Cors==4.0.1
grpcio==1.64.0
grpcio-testing==1.64.0
grpcio-tools==1.64.0
mod_wsgi==4.9.4
mysql-connector-python==8.0.29
mysqlclient==2.1.1
nest-asyncio==1.5.6
peewee==3.17.5
phonenumbers==8.13.36
pycryptodome==3.19.1
pytest==8.2.1
requests==2.32.2
SwobThirdPartyPlatforms @ git+https://github.com/smswithoutborders/SMSWithoutBorders-Customized-Third-Party-Platforms.git@main#egg=SwobThirdPartyPlatforms
twilio==7.12.0
Werkzeug==3.0.3
